src/main/webapp/app/entities/schedule/update/schedule-form.service.spec.ts
import { TestBed } from '@angular/core/testing';

import { sampleWithRequiredData, sampleWithNewData } from '../schedule.test-samples';

import { ScheduleFormService } from './schedule-form.service';

describe('Schedule Form Service', () => {
  let service: ScheduleFormService;

  beforeEach(() => {
    TestBed.configureTestingModule({});
    service = TestBed.inject(ScheduleFormService);
  });

  describe('Service methods', () => {
    describe('createScheduleFormGroup', () => {
      it('should create a new form with FormControl', () => {
        const formGroup = service.createScheduleFormGroup();

        expect(formGroup.controls).toEqual(
          expect.objectContaining({
            id: expect.any(Object),
            rule: expect.any(Object),
            firstInterval: expect.any(Object),
            secondInterval: expect.any(Object),
            thirdInterval: expect.any(Object),
            createdAt: expect.any(Object),
            updatedAt: expect.any(Object),
            agent: expect.any(Object),
          })
        );
      });

      it('passing ISchedule should create a new form with FormGroup', () => {
        const formGroup = service.createScheduleFormGroup(sampleWithRequiredData);

        expect(formGroup.controls).toEqual(
          expect.objectContaining({
            id: expect.any(Object),
            rule: expect.any(Object),
            firstInterval: expect.any(Object),
            secondInterval: expect.any(Object),
            thirdInterval: expect.any(Object),
            createdAt: expect.any(Object),
            updatedAt: expect.any(Object),
            agent: expect.any(Object),
          })
        );
      });
    });

    describe('getSchedule', () => {
      it('should return NewSchedule for default Schedule initial value', () => {
        // eslint-disable-next-line @typescript-eslint/no-unused-vars
        const formGroup = service.createScheduleFormGroup(sampleWithNewData);

        const schedule = service.getSchedule(formGroup) as any;

        expect(schedule).toMatchObject(sampleWithNewData);
      });

      it('should return NewSchedule for empty Schedule initial value', () => {
        const formGroup = service.createScheduleFormGroup();

        const schedule = service.getSchedule(formGroup) as any;

        expect(schedule).toMatchObject({});
      });

      it('should return ISchedule', () => {
        const formGroup = service.createScheduleFormGroup(sampleWithRequiredData);

        const schedule = service.getSchedule(formGroup) as any;

        expect(schedule).toMatchObject(sampleWithRequiredData);
      });
    });

    describe('resetForm', () => {
      it('passing ISchedule should not enable id FormControl', () => {
        const formGroup = service.createScheduleFormGroup();
        expect(formGroup.controls.id.disabled).toBe(true);

        service.resetForm(formGroup, sampleWithRequiredData);

        expect(formGroup.controls.id.disabled).toBe(true);
      });

      it('passing NewSchedule should disable id FormControl', () => {
        const formGroup = service.createScheduleFormGroup(sampleWithRequiredData);
        expect(formGroup.controls.id.disabled).toBe(true);

        service.resetForm(formGroup, { id: null });

        expect(formGroup.controls.id.disabled).toBe(true);
      });
    });
  });
});
